[New_Architecture] - Punto de guardo, faltan los diferentes actions y procesar la respuesta, no hay nada probado ni compilado
