+++ .idea/modules/es.wolfteam.wolfDiscordBot.test.iml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id="wolfDiscordBot:test" external.linked.project.path="$MODULE_DIR$/../.." external.root.project.path="$MODULE_DIR$/../.." external.system.id="GRADLE" external.system.module.group="es.wolfteam" external.system.module.type="sourceSet" external.system.module.version="1.0510-RELEASE" type="JAVA_MODULE" version="4">
+<module external.linked.project.id="wolfDiscordBot:test" external.linked.project.path="$MODULE_DIR$/../.." external.root.project.path="$MODULE_DIR$/../.." external.system.id="GRADLE" external.system.module.group="es.wolfteam" external.system.module.type="sourceSet" external.system.module.version="1.20190818-RELEASE" type="JAVA_MODULE" version="4">
   <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8">
     <output-test url="file://$MODULE_DIR$/../../out/test/classes" />
     <exclude-output />
