+++ src/main/java/es/wolfteam/core/WAction.java
@@ -1,0 +1,6 @@
+package es.wolfteam.core;
+
+public interface WAction
+{
+    void sendMessageEvent(final );
+}
