+++ src/main/java/es/wolfteam/exceptions/FilterException.java
@@ -1,0 +1,5 @@
+package es.wolfteam.exceptions;
+
+public class FilterException extends Throwable
+{
+}
