+++ src/main/java/es/wolfteam/filters/FilterMessage.java
@@ -1,0 +1,9 @@
+package es.wolfteam.filters;
+
+import es.wolfteam.data.FunctionData;
+import es.wolfteam.exceptions.FilterException;
+
+public interface FilterMessage
+{
+    FunctionData filterMessage(String message) throws FilterException;
+}
